+++ packages/models/src/models/LivechatVisitors.ts
@@ -380,18 +380,20 @@ export class LivechatVisitorsRaw extends BaseRaw<ILivechatVisitor> implements IL
 			.filter((phone) => phone?.trim().replace(/[^\d]/g, ''))
 			.map((phone) => ({ phoneNumber: phone }));
 
-		const update: UpdateFilter<ILivechatVisitor> = {
-			$addToSet: {
-				...(saveEmail.length && { visitorEmails: { $each: saveEmail } }),
-				...(savePhone.length && { phone: { $each: savePhone } }),
-			},
-		};
-
-		if (!Object.keys(update.$addToSet as Record<string, any>).length) {
+		if (!saveEmail.length && !savePhone.length) {
 			return Promise.resolve();
 		}
 
-		return this.updateOne({ _id }, update);
+		// the only reason we're using $setUnion here instead of $addToSet is because
+		// old visitors might have `visitorEmails` or `phone` as `null` which would cause $addToSet to fail
+		return this.updateOne({ _id }, [
+			{
+				$set: {
+					...(saveEmail.length && { visitorEmails: { $setUnion: [{ $ifNull: ['$visitorEmails', []] }, saveEmail] } }),
+					...(savePhone.length && { phone: { $setUnion: [{ $ifNull: ['$phone', []] }, savePhone] } }),
+				},
+			},
+		]);
 	}
 
 	removeContactManagerByUsername(manager: string): Promise<Document | UpdateResult> {