+++ packages/ui-voip/src/v2/components/PeerAutocomplete.tsx
@@ -1,4 +1,5 @@
-import { AutoComplete, Option, Avatar, Field, FieldRow, FieldDescription, FieldError } from '@rocket.chat/fuselage';
+import { UserStatus } from '@rocket.chat/core-typings';
+import { AutoComplete, Option, Avatar, Field, FieldRow, FieldDescription, FieldError, StatusBullet, Box } from '@rocket.chat/fuselage';
 import { useId } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -7,6 +8,7 @@ import { isFirstPeerAutocompleteOption } from '../MediaCallContext';
 export type PeerAutocompleteOptions = {
 	value: string; // user id
 	label: string; // name or username
+	status?: UserStatus;
 	identifier?: string | number; // extension number
 	avatarUrl?: string;
 };
@@ -44,7 +46,19 @@ const PeerAutocomplete = ({ options, filter, value, onChangeValue, onChangeFilte
 							return <Option key={value} label={label} icon='phone-out' {...props} />;
 						}
 						const thisOption = options.find((option) => option.value === value);
-						return <Option key={value} label={label} avatar={<Avatar size='x20' url={thisOption?.avatarUrl || ''} />} {...props} />;
+						return (
+							<Option
+								key={value}
+								label={
+									<Box display='flex' flexDirection='row' alignItems='center'>
+										<StatusBullet status={thisOption?.status} />
+										<Box mis={4}>{label}</Box>
+									</Box>
+								}
+								avatar={<Avatar size='x20' url={thisOption?.avatarUrl || ''} />}
+								{...props}
+							/>
+						);
 					}}
 					renderSelected={() => null}
 				/>