+++ packages/ui-voip/src/v2/components/PeerInfo/PeerInfo.stories.tsx
@@ -1,3 +1,4 @@
+import { UserStatus } from '@rocket.chat/core-typings';
 import type { Meta, StoryFn } from '@storybook/react';
 
 import { PeerInfo } from '.';
@@ -13,6 +14,7 @@ export const InternalUser: StoryFn<typeof PeerInfo> = () => {
 		<PeerInfo
 			callerId='1234'
 			displayName='John Doe'
+			status={UserStatus.ONLINE}
 			avatarUrl='data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wBDAAgGBgcGBQgHBwcJCQgKDBQNDAsLDBkSEw8UHRofHh0aHBwgJC
               4nICIsIxwcKDcpLDAxNDQ0Hyc5PTgyPC4zNDL/2wBDAQkJCQwLDBgNDRgyIRwhMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMj
               IyMjIyMjIyMjIyMjIyMjL/wAARCAAoACgDASIAAhEBAxEB/8QAGwAAAgIDAQAAAAAAAAAAAAAAAAcEBgIDBQj/xAAuEAACAQQAAwcEAQUAAA